requirements.txt
asgiref==3.7.2
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
rest-framework-simplejwt==0.0.2
python-decouple==3.8
gunicorn==21.2.0
packaging==23.2
drf_spectacular-0.27.1